Renamespace event worker event factories
