+++ tests/Functional/Services/WorkerEventFactory/JobTimeoutEventCallbackFactoryTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Services\EventCallbackFactory;
+namespace App\Tests\Functional\Services\WorkerEventFactory;
 
-use App\Services\EventCallbackFactory\EventCallbackFactoryInterface;
-use App\Services\EventCallbackFactory\JobTimeoutEventCallbackFactory;
+use App\Services\WorkerEventFactory\EventCallbackFactoryInterface;
+use App\Services\WorkerEventFactory\JobTimeoutEventCallbackFactory;
 use App\Tests\DataProvider\CallbackFactory\CreateFromJobTimeoutEventDataProviderTrait;
 
 class JobTimeoutEventCallbackFactoryTest extends AbstractEventCallbackFactoryTest
